pyproject.toml
[project]
name = "python-chess-board"
version = "0.1.0"
description = "A PySide6 chess board widget"
authors = [{name = "Leszek", email = "leszek.hanusz@gmail.com"}]
dependencies = [
    "PySide6",
    "python-chess",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "mypy",
    "pytest",
    "pytest-qt",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
python_chess_board = ["py.typed", "assets/*.svg"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
ignore_missing_imports = true
